requirements.txt
Flask
flask-cors
tensorflow
numpy
Pillow
google-generativeai
firebase-admin
google-cloud-storage
gunicorn
python-dotenv
geoip2==4.7.0
requests==2.31.0
